requirements.txt
asgiref==3.4.1
Deprecated==1.2.13
Django==4.0.1
django-cors-headers==3.10.1
djangorestframework==3.13.1
hashids==1.3.1
packaging==21.3
pyparsing==3.0.6
pytz==2021.3
redis==4.1.0
sqlparse==0.4.2
tzdata==2021.5
wrapt==1.13.3
